refactor(fdr): extract FDP bounds and weighted selection internals
Move FDP mathematics, WCS validation and pruning, and shared validators into private modules. Keep existing public APIs and statistical calculations unchanged.

Validated the isolated commit with 368 statistical, public API, and FDR integration tests, plus Ruff.
