+++ tests/unit/utils/stat/test_fdr_branch_coverage.py
@@ -2,16 +2,19 @@ import numpy as np
 import pytest
 
 from nonconform._internal import Pruning
-from nonconform.fdr import (
-    _as_1d_numeric,
+from nonconform._internal.validation import (
+    as_1d_numeric,
+    validate_finite,
+    validate_non_negative_finite,
+    validate_p_values,
+)
+from nonconform._internal.wcs import (
     _compute_r_star,
-    _compute_rejection_set_size_for_instance,
-    _extract_kde_support,
-    _run_wcs,
-    _validate_finite,
-    _validate_non_negative_finite,
-    _validate_p_values,
+    _rejection_set_size_for_instance,
     _validate_pruning,
+    _WCSInputs,
+    extract_kde_support,
+    run,
 )
 from nonconform.structures import ConformalResult
 
@@ -27,28 +30,33 @@ def _valid_arrays() -> tuple[
     return p_values, test_scores, calib_scores, test_weights, calib_weights
 
 
+def _test_inputs() -> _WCSInputs:
+    """Return a small validated-shape WCS bundle for helper tests."""
+    return _WCSInputs(*_valid_arrays())
+
+
 def test_as_1d_numeric_rejects_non_numeric_values() -> None:
     with pytest.raises(ValueError, match="numeric array"):
-        _as_1d_numeric("values", np.array(["x"], dtype=object))
+        as_1d_numeric("values", np.array(["x"], dtype=object))
 
 
 def test_validation_helpers_accept_empty_arrays() -> None:
     empty = np.array([], dtype=float)
-    _validate_non_negative_finite("weights", empty)
-    _validate_finite("scores", empty)
-    _validate_p_values(empty)
+    validate_non_negative_finite("weights", empty)
+    validate_finite("scores", empty)
+    validate_p_values(empty)
 
 
 def test_validate_non_negative_finite_rejects_non_finite_and_negative() -> None:
     with pytest.raises(ValueError, match="must be finite"):
-        _validate_non_negative_finite("weights", np.array([np.inf]))
+        validate_non_negative_finite("weights", np.array([np.inf]))
     with pytest.raises(ValueError, match="must be non-negative"):
-        _validate_non_negative_finite("weights", np.array([-1.0]))
+        validate_non_negative_finite("weights", np.array([-1.0]))
 
 
 def test_validate_finite_rejects_non_finite_values() -> None:
     with pytest.raises(ValueError, match="must be finite"):
-        _validate_finite("scores", np.array([np.nan]))
+        validate_finite("scores", np.array([np.nan]))
 
 
 def test_validate_pruning_rejects_wrong_type() -> None:
@@ -62,10 +70,9 @@ def test_compute_r_star_empty_metrics_returns_zero() -> None:
 
 def test_compute_rejection_set_size_requires_score_rank_cache() -> None:
     with pytest.raises(ValueError, match="missing score-rank cache"):
-        _compute_rejection_set_size_for_instance(
-            j=0,
-            test_scores=np.array([1.0, 2.0]),
-            w_test=np.array([1.0, 1.0]),
+        _rejection_set_size_for_instance(
+            index=0,
+            inputs=_test_inputs(),
             sum_calib_weight=2.0,
             bh_thresholds=np.array([0.05, 0.10]),
             calib_mass_strictly_above=np.array([1.0, 0.5]),
@@ -78,10 +85,9 @@ def test_compute_rejection_set_size_requires_score_rank_cache() -> None:
 
 def test_compute_rejection_set_size_rejects_non_positive_effective_mass() -> None:
     with pytest.raises(ValueError, match="positive finite effective calibration mass"):
-        _compute_rejection_set_size_for_instance(
-            j=0,
-            test_scores=np.array([1.0, 2.0]),
-            w_test=np.array([1.0, 1.0]),
+        _rejection_set_size_for_instance(
+            index=0,
+            inputs=_test_inputs(),
             sum_calib_weight=0.0,
             bh_thresholds=np.array([0.05, 0.10]),
             calib_mass_strictly_above=np.array([0.0, 0.0]),
@@ -95,7 +101,7 @@ def test_compute_rejection_set_size_rejects_non_positive_effective_mass() -> Non
 def test_extract_kde_support_requires_dict_metadata() -> None:
     result = ConformalResult(metadata={"kde": 1})
     with pytest.raises(ValueError, match="must be a dictionary"):
-        _extract_kde_support(result)
+        extract_kde_support(result)
 
 
 def test_extract_kde_support_rejects_non_numeric_total_weight() -> None:
@@ -109,7 +115,7 @@ def test_extract_kde_support_rejects_non_numeric_total_weight() -> None:
         }
     )
     with pytest.raises(ValueError, match="finite positive float"):
-        _extract_kde_support(result)
+        extract_kde_support(result)
 
 
 def test_extract_kde_support_rejects_too_short_eval_grid() -> None:
@@ -123,7 +129,7 @@ def test_extract_kde_support_rejects_too_short_eval_grid() -> None:
         }
     )
     with pytest.raises(ValueError, match="at least 2 points"):
-        _extract_kde_support(result)
+        extract_kde_support(result)
 
 
 def test_extract_kde_support_rejects_length_mismatch() -> None:
@@ -137,7 +143,7 @@ def test_extract_kde_support_rejects_length_mismatch() -> None:
         }
     )
     with pytest.raises(ValueError, match="must have equal length"):
-        _extract_kde_support(result)
+        extract_kde_support(result)
 
 
 def test_extract_kde_support_rejects_cdf_out_of_bounds() -> None:
@@ -151,13 +157,13 @@ def test_extract_kde_support_rejects_cdf_out_of_bounds() -> None:
         }
     )
     with pytest.raises(ValueError, match=r"within \[0, 1\]"):
-        _extract_kde_support(result)
+        extract_kde_support(result)
 
 
 def test_run_wcs_rejects_mismatched_test_lengths() -> None:
     p_values, test_scores, calib_scores, _, calib_weights = _valid_arrays()
     with pytest.raises(ValueError, match="must have the same length"):
-        _run_wcs(
+        run(
             p_values=p_values,
             test_scores=test_scores,
             calib_scores=calib_scores,
@@ -172,7 +178,7 @@ def test_run_wcs_rejects_mismatched_test_lengths() -> None:
 def test_run_wcs_rejects_mismatched_calibration_lengths() -> None:
     p_values, test_scores, calib_scores, test_weights, _ = _valid_arrays()
     with pytest.raises(ValueError, match="must have the same length"):
-        _run_wcs(
+        run(
             p_values=p_values,
             test_scores=test_scores,
             calib_scores=calib_scores,
@@ -187,7 +193,7 @@ def test_run_wcs_rejects_mismatched_calibration_lengths() -> None:
 def test_run_wcs_rejects_non_positive_total_calibration_weight() -> None:
     p_values, test_scores, calib_scores, test_weights, _ = _valid_arrays()
     with pytest.raises(ValueError, match="positive finite total calibration weight"):
-        _run_wcs(
+        run(
             p_values=p_values,
             test_scores=test_scores,
             calib_scores=calib_scores,
